masive refactoring
https://github.com/metasfresh/metasfresh-webui-api/issues/668
