+++ src/main/java/de/metas/ui/web/handlingunits/HUEditorViewBuffer_HighVolume.java
@@ -1,11 +1,7 @@
 package de.metas.ui.web.handlingunits;
 
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.Collection;
 import java.util.Iterator;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
@@ -13,15 +9,9 @@ import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.UnaryOperator;
 import java.util.stream.Stream;
 
-import org.adempiere.ad.expression.api.IExpressionEvaluator.OnVariableNotFound;
-import org.adempiere.ad.expression.api.IStringExpression;
-import org.adempiere.ad.trx.api.ITrx;
-import org.adempiere.exceptions.DBException;
 import org.adempiere.util.collections.IteratorUtils;
 import org.adempiere.util.collections.PagedIterator.PageFetcher;
 import org.compiere.util.CCache;
-import org.compiere.util.DB;
-import org.compiere.util.Env;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
@@ -29,13 +19,9 @@ import com.google.common.collect.Iterables;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.ui.web.document.filter.DocumentFilter;
 import de.metas.ui.web.exceptions.EntityNotFoundException;
-import de.metas.ui.web.view.SqlViewRowIdsOrderedSelectionFactory;
-import de.metas.ui.web.view.ViewEvaluationCtx;
+import de.metas.ui.web.handlingunits.HUIdsFilterHelper.HUIdsFilterData;
 import de.metas.ui.web.view.ViewId;
 import de.metas.ui.web.view.ViewRowIdsOrderedSelection;
-import de.metas.ui.web.view.descriptor.SqlViewBinding;
-import de.metas.ui.web.view.descriptor.SqlViewRowIdsConverter;
-import de.metas.ui.web.view.descriptor.SqlViewSelectionQueryBuilder;
 import de.metas.ui.web.window.datatypes.DocumentId;
 import de.metas.ui.web.window.datatypes.DocumentIdsSelection;
 import de.metas.ui.web.window.model.DocumentQueryOrderBy;
@@ -65,18 +51,12 @@ import lombok.NonNull;
 
 public class HUEditorViewBuffer_HighVolume implements HUEditorViewBuffer
 {
-	public static final int HIGHVOLUME_THRESHOLD = 100;
-
-	private static final int STREAM_ALL_MAX_SIZE_ALLOWED = 100;
+	private static final int HIGHVOLUME_THRESHOLD = 100;
+	private static final int STREAM_ALL_MAX_SIZE_ALLOWED = 200;
 
 	private final HUEditorViewRepository huEditorRepo;
-	private final IStringExpression sqlSelectHUIdsByPage;
-
 	private final ImmutableList<DocumentFilter> stickyFilters;
-
-	private final SqlViewRowIdsOrderedSelectionFactory viewSelectionFactory;
 	private final AtomicReference<ViewRowIdsOrderedSelection> defaultSelectionRef;
-
 	private final CCache<DocumentId, HUEditorRow> cache_huRowsById = CCache.newLRUCache(I_M_HU.Table_Name + "#HUEditorRows#by#Id", 100, 2);
 
 	HUEditorViewBuffer_HighVolume(
@@ -86,16 +66,10 @@ public class HUEditorViewBuffer_HighVolume implements HUEditorViewBuffer
 			final List<DocumentFilter> filters)
 	{
 		this.huEditorRepo = huEditorRepo;
-
-		final SqlViewBinding viewBinding = huEditorRepo.getSqlViewBinding();
-		viewSelectionFactory = SqlViewRowIdsOrderedSelectionFactory.of(viewBinding);
-		sqlSelectHUIdsByPage = viewBinding.getSqlSelectByPage();
-
 		this.stickyFilters = ImmutableList.copyOf(stickyFilters);
 
-		final ViewEvaluationCtx viewEvalCtx = ViewEvaluationCtx.of(Env.getCtx());
 		final List<DocumentFilter> filtersAll = ImmutableList.copyOf(Iterables.concat(stickyFilters, filters));
-		final ViewRowIdsOrderedSelection defaultSelection = viewSelectionFactory.createOrderedSelection(viewEvalCtx, viewId, filtersAll, viewBinding.getDefaultOrderBys());
+		final ViewRowIdsOrderedSelection defaultSelection = huEditorRepo.createDefaultSelection(viewId, filtersAll);
 		defaultSelectionRef = new AtomicReference<>(defaultSelection);
 	}
 
@@ -103,11 +77,6 @@ public class HUEditorViewBuffer_HighVolume implements HUEditorViewBuffer
 	public List<DocumentFilter> getStickyFilters()
 	{
 		return stickyFilters;
-	}
-
-	private SqlViewRowIdsConverter getRowIdsConverter()
-	{
-		return huEditorRepo.getRowIdsConverter();
 	}
 
 	private ViewRowIdsOrderedSelection getDefaultSelection()
@@ -138,6 +107,7 @@ public class HUEditorViewBuffer_HighVolume implements HUEditorViewBuffer
 	@Override
 	public void invalidateAll()
 	{
+		huEditorRepo.invalidateCache();
 		cache_huRowsById.clear();
 	}
 
@@ -155,7 +125,7 @@ public class HUEditorViewBuffer_HighVolume implements HUEditorViewBuffer
 			return false;
 		}
 
-		return changeDefaultSelection(defaultSelection -> viewSelectionFactory.addRowIdsToSelection(defaultSelection, rowIdsToAdd));
+		return changeDefaultSelection(defaultSelection -> huEditorRepo.addRowIdsToSelection(defaultSelection, rowIdsToAdd));
 	}
 
 	@Override
@@ -170,14 +140,14 @@ public class HUEditorViewBuffer_HighVolume implements HUEditorViewBuffer
 
 		rowIdsToRemove.forEach(rowId -> cache_huRowsById.remove(rowId));
 
-		return changeDefaultSelection(defaultSelection -> viewSelectionFactory.removeRowIdsFromSelection(defaultSelection, rowIdsToRemove));
+		return changeDefaultSelection(defaultSelection -> huEditorRepo.removeRowIdsFromSelection(defaultSelection, rowIdsToRemove));
 	}
 
 	@Override
 	public boolean containsAnyOfHUIds(final Collection<Integer> huIdsToCheck)
 	{
 		final DocumentIdsSelection rowIds = HUEditorRowId.rowIdsFromTopLevelM_HU_IDs(huIdsToCheck);
-		return viewSelectionFactory.containsAnyOfRowIds(getDefaultSelection().getSelectionId(), rowIds);
+		return huEditorRepo.containsAnyOfRowIds(getDefaultSelection(), rowIds);
 	}
 
 	@Override
@@ -269,50 +239,7 @@ public class HUEditorViewBuffer_HighVolume implements HUEditorViewBuffer
 	private PageFetcher<Integer> huIdsPageFetcher(final List<DocumentQueryOrderBy> orderBys)
 	{
 		final ViewRowIdsOrderedSelection selection = getSelection(orderBys);
-		final String viewSelectionId = selection.getSelectionId();
-		final ViewEvaluationCtx viewEvalCtx = ViewEvaluationCtx.of(Env.getCtx());
-		final String sql = sqlSelectHUIdsByPage.evaluate(viewEvalCtx.toEvaluatee(), OnVariableNotFound.Fail);
-
-		return (firstRow, maxRows) -> {
-			final int firstSeqNo = firstRow + 1; // NOTE: firstRow is 0-based while SeqNo are 1-based
-			final int lastSeqNo = firstRow + maxRows;
-
-			final Object[] sqlParams = new Object[] { viewSelectionId, firstSeqNo, lastSeqNo };
-
-			PreparedStatement pstmt = null;
-			ResultSet rs = null;
-			try
-			{
-				pstmt = DB.prepareStatement(sql, ITrx.TRXNAME_ThreadInherited);
-				pstmt.setMaxRows(maxRows);
-				DB.setParameters(pstmt, sqlParams);
-
-				rs = pstmt.executeQuery();
-
-				final Set<Integer> huIds = new LinkedHashSet<>();
-				while (rs.next())
-				{
-					final int huId = rs.getInt(I_M_HU.COLUMNNAME_M_HU_ID);
-					if (huId <= 0)
-					{
-						continue;
-					}
-					huIds.add(huId);
-				}
-
-				return huIds;
-
-			}
-			catch (final SQLException ex)
-			{
-				throw DBException.wrapIfNeeded(ex)
-						.setSqlIfAbsent(sql, sqlParams);
-			}
-			finally
-			{
-				DB.close(rs, pstmt);
-			}
-		};
+		return (firstRow, maxRows) -> huEditorRepo.retrievePagedHUIds(selection, firstRow, maxRows);
 	}
 
 	private Stream<Integer> streamHUIdsByPage(final int firstRow, final int maxRows, final List<DocumentQueryOrderBy> orderBys)
@@ -347,8 +274,33 @@ public class HUEditorViewBuffer_HighVolume implements HUEditorViewBuffer
 	@Override
 	public String getSqlWhereClause(final DocumentIdsSelection rowIds)
 	{
-		final String sqlKeyColumnNameFK = I_M_HU.Table_Name + "." + I_M_HU.COLUMNNAME_M_HU_ID;
-		final String selectionId = getViewId().getViewId();
-		return SqlViewSelectionQueryBuilder.buildSqlWhereClause(sqlKeyColumnNameFK, selectionId, rowIds, getRowIdsConverter());
+		return huEditorRepo.buildSqlWhereClause(getDefaultSelection(), rowIds);
 	}
+
+	public static boolean isHighVolume(final List<DocumentFilter> stickyFilters)
+	{
+		final HUIdsFilterData huIdsFilterData = HUIdsFilterHelper.extractFilterDataOrNull(stickyFilters);
+		if (huIdsFilterData == null)
+		{
+			return true;
+		}
+
+		final Set<Integer> huIds = huIdsFilterData.getInitialHUIds();
+		if (huIds == null)
+		{
+			// null means no restrictions, so we might have a lot of HUs
+			return true; // high volume
+		}
+		else if (huIds.isEmpty())
+		{
+			// no HUs will be allowed
+			return false; // not high volume
+		}
+		else
+		{
+			// consider high volume if it's above give threshold
+			return huIds.size() >= HIGHVOLUME_THRESHOLD;
+		}
+	}
+
 }
