+++ src/main/java/de/metas/ui/web/view/ViewRowIdsOrderedSelectionFactory.java
@@ -42,4 +42,6 @@ public interface ViewRowIdsOrderedSelectionFactory
 	ViewRowIdsOrderedSelection addRowIdsToSelection(ViewRowIdsOrderedSelection selection, DocumentIdsSelection rowIds);
 
 	ViewRowIdsOrderedSelection removeRowIdsFromSelection(ViewRowIdsOrderedSelection selection, DocumentIdsSelection rowIds);
+
+	boolean containsAnyOfRowIds(ViewRowIdsOrderedSelection selection, DocumentIdsSelection rowIds);
 }
